block detection
issue 1) this version of the file results in too many false positives. this can be resolved by checking for the presence of yellow alongside the bounding boxes. issue 2) the program seems to be too resource intensive, resulting in crashes. additionally, make sure to release the matricies that are no longer in use. (disclaimer: from pure observation,) it appears that setting a region of interest to crop the matrix in is also a cause of crashes, as the resolution depends on the phone orientation and model. checking for the presence of yellow alongside black areas will eliminiate the need for cropping.